Put everything into single laffy package
