+++ src/main/java/laffy/AddEventCommand.java
@@ -1,3 +1,5 @@
+package laffy;
+
 public class AddEventCommand extends Command {
     public static final String COMMAND_WORD = "event";
     private String desc;
