+++ src/main/java/laffy/AddTodoCommand.java
@@ -1,3 +1,5 @@
+package laffy;
+
 public class AddTodoCommand extends Command {
     public static final String COMMAND_WORD = "todo";
     private String desc;
