+++ src/main/java/laffy/DeleteCommand.java
@@ -1,7 +1,9 @@
-class UnmarkCommand extends IndexedCommand {
-    public static final String COMMAND_WORD = "unmark";
+package laffy;
 
-    public UnmarkCommand(String args) {
+public class DeleteCommand extends IndexedCommand {
+    public static final String COMMAND_WORD = "delete";
+
+    public DeleteCommand(String args) {
         super(args);
     }
 
@@ -10,7 +12,7 @@ class UnmarkCommand extends IndexedCommand {
         if (!isValid) {
             return "Index must be an integer.\n" + getUsage();
         } else {
-            String result = taskList.markAsUndone(this.index);
+            String result = taskList.delete(this.index);
             super.execute(taskList, ui, storage);
             return result;
         }
