+++ src/main/java/laffy/Event.java
@@ -1,3 +1,5 @@
+package laffy;
+
 import java.util.ArrayList;
 import java.time.LocalDateTime;
