+++ src/main/java/laffy/ExitCommand.java
@@ -1,3 +1,5 @@
+package laffy;
+
 public class ExitCommand extends Command {
     public static final String COMMAND_WORD = "bye";
