+++ src/main/java/laffy/IndexedCommand.java
@@ -1,3 +1,5 @@
+package laffy;
+
 public class IndexedCommand extends Command {
     protected final int index;
     public static final String COMMAND_WORD = "indexed";
