+++ src/main/java/laffy/InvalidCommand.java
@@ -1,3 +1,5 @@
+package laffy;
+
 public class InvalidCommand extends Command {
     public static final String COMMAND_WORD = "invalid";
