+++ src/main/java/laffy/Laffy.java
@@ -1,3 +1,5 @@
+package laffy;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
