+++ src/main/java/laffy/ListCommand.java
@@ -1,3 +1,5 @@
+package laffy;
+
 public class ListCommand extends Command {
     public static final String COMMAND_WORD = "list";
