+++ src/main/java/laffy/MarkCommand.java
@@ -1,3 +1,5 @@
+package laffy;
+
 public class MarkCommand extends IndexedCommand {
     public static final String COMMAND_WORD = "mark";
