+++ src/main/java/laffy/Storage.java
@@ -1,3 +1,5 @@
+package laffy;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
