+++ src/main/java/laffy/Task.java
@@ -1,3 +1,5 @@
+package laffy;
+
 import java.util.ArrayList;
 
 public class Task {
