+++ src/main/java/laffy/Ui.java
@@ -1,3 +1,5 @@
+package laffy;
+
 import java.util.Scanner;
 
 public class Ui {
